Add Wagtail to our blog

+++ blog/views.py
@@ -1,20 +1,13 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
+from .models import Post, Comment
+from .forms import PostForm, CommentForm
 from django.contrib.auth.decorators import login_required
-from .models import Post, Comment  # 🔹 импортировали Comment
-from .forms import PostForm, CommentForm  # 🔹 импортировали CommentForm
 
-# Список опубликованных постов
 def post_list(request):
-    posts = Post.objects.filter(published_date__isnull=False).order_by('published_date')
+    posts = Post.objects.filter(published_date__isnull=False).order_by('-published_date')
     return render(request, 'blog/post_list.html', {'posts': posts})
 
-# Детали конкретного поста
-def post_detail(request, pk):
-    post = get_object_or_404(Post, pk=pk)
-    return render(request, 'blog/post_detail.html', {'post': post})
-
-# Создание нового поста
 @login_required
 def post_new(request):
     if request.method == "POST":
@@ -22,13 +15,13 @@ def post_new(request):
         if form.is_valid():
             post = form.save(commit=False)
             post.author = request.user
+            # сохраняем как черновик (не ставим published_date)
             post.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm()
     return render(request, 'blog/post_edit.html', {'form': form})
 
-# Редактирование поста
 @login_required
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
@@ -43,28 +36,29 @@ def post_edit(request, pk):
         form = PostForm(instance=post)
     return render(request, 'blog/post_edit.html', {'form': form})
 
-# Список черновиков
+def post_detail(request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    return render(request, 'blog/post_detail.html', {'post': post})
+
 @login_required
 def post_draft_list(request):
     posts = Post.objects.filter(published_date__isnull=True).order_by('created_date')
     return render(request, 'blog/post_draft_list.html', {'posts': posts})
 
-# Публикация поста
 @login_required
 def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.published_date = timezone.now()
-    post.save()
+    if request.method == 'POST':
+        post.publish()
     return redirect('post_detail', pk=pk)
 
-# Удаление поста
 @login_required
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
+    if request.method == 'POST':
+        post.delete()
     return redirect('post_list')
 
-# 🔹 Добавление комментариев к посту
 def add_comment_to_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
     if request.method == "POST":
@@ -78,18 +72,18 @@ def add_comment_to_post(request, pk):
         form = CommentForm()
     return render(request, 'blog/add_comment_to_post.html', {'form': form})
 
-# 🔹 Одобрение комментария
 @login_required
 def comment_approve(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.approve()
+    if request.method == 'POST':
+        comment.approve()
     return redirect('post_detail', pk=comment.post.pk)
 
-# 🔹 Удаление комментария
 @login_required
 def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.delete()
+    if request.method == 'POST':
+        comment.delete()
     return redirect('post_detail', pk=comment.post.pk)
